--- dev/benchmarks/.gitignore
@@ -1,1 +1,0 @@
-mega_gallery/
