--- dev/tools/.gitignore
@@ -1,3 +1,0 @@
-.packages
-pubspec.lock
-packages
