--- dev/tools/pubspec.yaml
@@ -1,6 +1,0 @@
-name: dev_tools
-description: Various repository development tools for flutter.
-
-dependencies:
-  args: ^0.13.4
-  path: ^1.3.0
